requirements.txt
click==7.1.2
PyDrive==1.3.1
python_daemon==2.3.0
websocket_client==1.3.1
